acme_api/pyproject.toml
[project]
name = "acme-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-jose[cryptography]>=3.3",
  "httpx>=0.27",
  "pydantic>=2.8",
  "pydantic-settings>=2.0",
  "cachetools>=5.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
